# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Barbossa API
  description: |
    Family music library management API.

    ## Authentication
    All endpoints except `/api/auth/login` require a valid JWT token.
    Pass token in Authorization header: `Bearer <token>`

    ## User Roles
    - **Regular**: Can browse, heart, download, export own library
    - **Admin**: Can manage users, delete from disk, upload to TorrentLeech
  version: 0.1.9
  contact:
    name: Barbossa
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://barbossa:8080
    description: Docker internal
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Logout
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
  /api/auth/me:
    get:
      tags: [auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==========================================================================
  # ARTISTS
  # ==========================================================================
...
  /api/albums/{id}/artwork:
    get:
      tags: [albums]
      summary: Get album artwork
      operationId: getAlbumArtwork
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album artwork
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
    put:
      tags: [albums]
      summary: Update album artwork
      operationId: updateAlbumArtwork
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                artwork:
                  type: string
                  format: binary
      responses:
        '200':
          description: Artwork updated
        '400':
          $ref: '#/components/responses/BadRequest'

  # ==========================================================================
  # TRACKS
  # ==========================================================================
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: List users
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin]
      summary: Create user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/admin/users/{id}:
    put:
      tags: [admin]
      summary: Update user
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
    delete:
      tags: [admin]
      summary: Delete user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
  /api/admin/rescan:
    post:
      tags: [admin]
      summary: Rescan library
      operationId: rescanLibrary
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Rescan started
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AlbumId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden (admin only)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        is_admin:
          type: boolean
          default: false
    UpdateUserRequest:
      type: object
      properties:
        password:
          type: string
        is_admin:
          type: boolean
...